Source/LibraryManagementSystem.Infrastructure.IntegrationTests/Implementations/BookDalTest.cs
using LibraryManagementSystem.Domain.Models;
using LibraryManagementSystem.Infrastructure.Implementations;
using LibraryManagementSystem.Infrastructure.IntegrationTests.Common.Helpers;
using System.Data.SqlClient;
using Xunit.Abstractions;

namespace LibraryManagementSystem.Infrastructure.IntegrationTests.Implementations
{
    [Collection(nameof(CommonHelper))]
    public class BookDalTest
    {
        private readonly ITestOutputHelper _outputHelper;
        private readonly CommonHelper _commonHelper;
        private readonly BookDal _bookDal;

        public BookDalTest(ITestOutputHelper outputHelper, CommonHelper commonHelper)
        {
            _outputHelper = outputHelper;
            _commonHelper = commonHelper;
            _bookDal = new BookDal(commonHelper.Settings.Database.ConnectionString);
        }

        /// <summary>
        /// Get top {x} amount of <see cref="Book"/>s.
        /// </summary>
        [Theory]
        [InlineData(3)]
        [InlineData(5)]
        public async Task GetTopBorrowedBooks(int numberOfTopBooks)
        {
            // Arrange
            List<TopBookDetail> topBorrowedBooks;

            await using var sqlConnection = new SqlConnection(_commonHelper.Settings.Database.ConnectionString);
            await sqlConnection.OpenAsync();
            var sqlCommand = sqlConnection.CreateCommand();
            var sqlTransaction = sqlConnection.BeginTransaction();
            sqlCommand.Transaction = sqlTransaction;

            await CommonHelper.AddTestDataToDatabase(sqlCommand);

            // Act
            try
            {
                topBorrowedBooks = await _bookDal.GetTopBorrowedBooks(sqlCommand, numberOfTopBooks);
            }
            finally
            {
                await sqlTransaction.RollbackAsync();
            }

            // Assert
            Assert.Equal(numberOfTopBooks, topBorrowedBooks.Count);

            foreach (var topBorrowedBook in topBorrowedBooks)
            {
                _outputHelper.WriteLine($"One of the top books is: {topBorrowedBook.Title}.");
            }
        }
    }
}